genie_cache_arch/requirements.txt
# Core Dependencies
databricks-sdk>=0.20.0

# Question Normalization (Multi-Agent - Optional)
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0

# Korean NLP (Simple Normalizer Phase 2)
kiwipiepy>=0.15.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data Models
pydantic>=2.5.0

# Retry Logic
tenacity>=8.2.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Testing (Dev only)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
